requirements.txt
altair==6.0.0
asttokens==3.0.0
attrs==26.1.0
blinker==1.9.0
cachetools==7.0.5
certifi==2026.2.25
charset-normalizer==3.4.7
click==8.1.8
colorama==0.4.6
comm==0.2.2
debugpy==1.8.13
decorator==5.2.1
executing==2.2.0
gitdb==4.0.12
GitPython==3.1.46
gTTS==2.5.4
idna==3.11
ipykernel==6.29.5
ipython==9.0.1
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter_client==8.6.3
jupyter_core==5.7.2
MarkupSafe==3.0.3
matplotlib-inline==0.1.7
narwhals==2.19.0
nest-asyncio==1.6.0
numpy==2.4.4
packaging==24.2
pandas==3.0.2
parso==0.8.4
pillow==12.2.0
platformdirs==4.3.6
prompt_toolkit==3.0.50
protobuf==7.34.1
psutil==7.0.0
pure_eval==0.2.3
pyarrow==23.0.1
pydeck==0.9.1
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pywin32==308
pyzmq==26.2.1
referencing==0.37.0
requests==2.33.1
rpds-py==0.30.0
six==1.17.0
smmap==5.0.3
stack-data==0.6.3
streamlit==1.56.0
tenacity==9.1.4
toml==0.10.2
tornado==6.4.2
traitlets==5.14.3
typing_extensions==4.15.0
tzdata==2026.1
urllib3==2.6.3
watchdog==6.0.0
wcwidth==0.2.13
